+++ ios/SmartShoppingList/SmartShoppingListTests/TestTags.swift
@@ -1,0 +1,6 @@
+import Testing
+
+extension Tag {
+    @Tag static var fast: Self
+    @Tag static var integration: Self
+}